+++ app/src/main/java/com/example/collesure2/ui/MainActivity.kt
@@ -1,30 +1,15 @@
 package com.example.collesure2.ui
 
-import android.app.ListFragment
-import android.content.Intent
-import android.os.AsyncTask
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Button
-import androidx.fragment.app.FragmentManager
 import com.example.collesure2.R
-import com.example.collesure2.ui.list.ResultFragment
-import kotlinx.android.synthetic.main.activity_main.*
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import java.util.regex.Pattern
-import androidx.core.app.ComponentActivity.ExtraData
-import androidx.core.content.ContextCompat.getSystemService
-import android.icu.lang.UCharacter.GraphemeClusterBreak.T
-import com.google.android.material.bottomnavigation.BottomNavigationView
-
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        buttom_navigation_view
+
     }
 
 }
