+++ app/src/main/java/com/example/collesure2/ui/list/RecyclerFragment.kt
@@ -1,0 +1,41 @@
+package com.example.collesure2.ui.list
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.example.collesure2.R
+
+
+
+class RecyclerFragment : Fragment() {
+
+    companion object {
+        fun newInstance(): RecyclerFragment {
+            val fragment = RecyclerFragment()
+            return fragment
+        }
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val bundle = arguments
+        val imageUrlList = bundle!!.get("imageUrlList") as List<String>
+
+        val recycler = view.findViewById<RecyclerView>(R.id.recyclerview).apply {
+            setHasFixedSize(true)
+            layoutManager = GridLayoutManager(context, 2)
+            adapter = RecyclerAdapter(context, imageUrlList)
+        }
+    }
+
+    override fun onCreateView(inflater: LayoutInflater,container: ViewGroup?,savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_recycler,container,false)
+    }
+
+}
