+++ app/src/main/java/com/example/collesure2/ui/user/HistoryFragment.kt
@@ -1,0 +1,4 @@
+package com.example.collesure2.ui.user
+
+class HistoryFragment {
+}
