+++ app/src/main/java/com/example/collesure2/ui/user/SearchFragment.kt
@@ -1,0 +1,13 @@
+package com.example.collesure2.ui.user
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+class SearchFragment : Fragment() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+//    override fun onCreateView(inflater: LayoutInflater,container: ViewGroup?,savedInstanceState: Bundle?): ? {
+//        return inflater.inflate(R.layout.activity_result,container,false)
+//    }
+}
