+++ app/src/main/java/com/example/collesure2/ui/user/SettingFragment.kt
@@ -1,0 +1,4 @@
+package com.example.collesure2.ui.user
+
+class SettingFragment {
+}
