+++ app/src/main/java/com/example/collesure2/ui/user/ViewActivity.kt
@@ -1,5 +1,4 @@
 package com.example.collesure2.ui.user
 
-class SeachFragment {
-
+class ViewActivity {
 }
